refactor: seperate user create logic from auth service
