+++ backend/src/auth/services/auth/auth.interface.ts
@@ -4,13 +4,13 @@ export interface IAuthService {
   defaultCookieOptions;
   accessTokenCookieOptions;
   refreshTokenCookieOptions;
-  validateUser(userDetails: UserDetails);
-  createUser(userDetails: UserDetails);
+  // validateUser(userDetails: UserDetails);
+  // createUser(userDetails: UserDetails);
   getAccessToken(id: number, two_factor_activated: boolean);
   getRefreshToken(id: number);
   setRefreshToken(id: number, refreshToken: string);
   verify(accessToken: any);
-  createDummyUser();
-  deleteDummyUser(user);
-  updateUserStatus(userId: number, status: Status);
+  // createDummyUser();
+  // deleteDummyUser(user);
+  // updateUserStatus(userId: number, status: Status);
 }
