+++ backend/src/auth/services/auth/auth.service.ts
@@ -42,20 +42,20 @@ export class AuthService implements IAuthService {
     return bcrypt.hash(data, 12);
   }
 
-  async validateUser(userDetails: UserDetails) {
-    // console.log('validateUser');
-    const { intra_id } = userDetails;
-    const user = await this.userRepository.findOneBy({ intra_id });
-    // console.log(user);
-    if (user) return user;
-    return this.createUser(userDetails);
-  }
+  // async validateUser(userDetails: UserDetails) {
+  //   // console.log('validateUser');
+  //   const { intra_id } = userDetails;
+  //   const user = await this.userRepository.findOneBy({ intra_id });
+  //   // console.log(user);
+  //   if (user) return user;
+  //   return this.createUser(userDetails);
+  // }
 
-  createUser(userDetails: UserDetails) {
-    console.log('creating a new user');
-    const user = this.userRepository.create(userDetails);
-    return this.userRepository.save(user);
-  }
+  // createUser(userDetails: UserDetails) {
+  //   console.log('creating a new user');
+  //   const user = this.userRepository.create(userDetails);
+  //   return this.userRepository.save(user);
+  // }
 
   getAccessToken(id: number, two_factor_activated: boolean) {
     // console.log('getAccessToken() two_factor_activated:', two_factor_activated);
@@ -113,81 +113,81 @@ export class AuthService implements IAuthService {
     }
   }
 
-  async validateDummy(userDetails: UserDetails) {
-    const { intra_id } = userDetails;
-    console.log('dummy intra id:', intra_id);
-    // const user = await this.userRepository.findOneBy({ intra_id });
-    const user = await this.userRepository
-      .createQueryBuilder('users')
-      .where('users.intra_id=:intra_id', { intra_id })
-      .getOne();
-    // console.log('dummyuser:', user);
-    return user;
-  }
+  //   async validateDummy(userDetails: UserDetails) {
+  //     const { intra_id } = userDetails;
+  //     console.log('dummy intra id:', intra_id);
+  //     // const user = await this.userRepository.findOneBy({ intra_id });
+  //     const user = await this.userRepository
+  //       .createQueryBuilder('users')
+  //       .where('users.intra_id=:intra_id', { intra_id })
+  //       .getOne();
+  //     // console.log('dummyuser:', user);
+  //     return user;
+  //   }
 
-  async createDummy(userDetails: UserDetails) {
-    const user = this.userRepository.create(userDetails);
-    return await this.userRepository.save(user);
-  }
+  //   async createDummy(userDetails: UserDetails) {
+  //     const user = this.userRepository.create(userDetails);
+  //     return await this.userRepository.save(user);
+  //   }
 
-  async createDummyUser() {
-    // console.log('createDummyUser()');
-    let name = 'dummy';
-    let userDetails = {
-      intra_id: name,
-      email: name,
-      image_url: process.env.DUMMY_URL,
-      username: name,
-    };
-    let dummy = await this.validateDummy(userDetails);
-    while (dummy) {
-      name = dummy.username.substring(0, 5);
-      let number = Number(dummy.username.replace(/[^0-9]/g, ''));
-      number += 1;
-      name += number;
-      userDetails = {
-        intra_id: name,
-        email: name,
-        image_url: process.env.DUMMY_URL,
-        username: name,
-      };
-      dummy = await this.validateDummy(userDetails);
-    }
-    // console.log('userdetails af if:', userDetails);
-    return this.createDummy(userDetails);
-  }
+  //   async createDummyUser() {
+  //     // console.log('createDummyUser()');
+  //     let name = 'dummy';
+  //     let userDetails = {
+  //       intra_id: name,
+  //       email: name,
+  //       image_url: process.env.DUMMY_URL,
+  //       username: name,
+  //     };
+  //     let dummy = await this.validateDummy(userDetails);
+  //     while (dummy) {
+  //       name = dummy.username.substring(0, 5);
+  //       let number = Number(dummy.username.replace(/[^0-9]/g, ''));
+  //       number += 1;
+  //       name += number;
+  //       userDetails = {
+  //         intra_id: name,
+  //         email: name,
+  //         image_url: process.env.DUMMY_URL,
+  //         username: name,
+  //       };
+  //       dummy = await this.validateDummy(userDetails);
+  //     }
+  //     // console.log('userdetails af if:', userDetails);
+  //     return this.createDummy(userDetails);
+  //   }
 
-  async deleteDummyUser(user) {
-    let count = user.intra_id.indexOf('dummy');
-    // console.log('deleteDummyUser:', count);
-    if (count === 0) {
-      // console.log('dummy:', user.intra_id);
-      const dummy = await this.userRepository
-        .createQueryBuilder('users')
-        .where('users.user_id=:id', { id: user.id })
-        .getOne();
-      await this.userRepository.remove(dummy);
-      return true;
-    }
-    return false;
-  }
+  //   async deleteDummyUser(user) {
+  //     let count = user.intra_id.indexOf('dummy');
+  //     // console.log('deleteDummyUser:', count);
+  //     if (count === 0) {
+  //       // console.log('dummy:', user.intra_id);
+  //       const dummy = await this.userRepository
+  //         .createQueryBuilder('users')
+  //         .where('users.user_id=:id', { id: user.id })
+  //         .getOne();
+  //       await this.userRepository.remove(dummy);
+  //       return true;
+  //     }
+  //     return false;
+  //   }
 
-  async updateUserStatus(userId: number, status: Status) {
-    const user = await this.userRepository
-      .createQueryBuilder('users')
-      .where('users.user_id=:userId', { userId })
-      .getOne();
-    if (!user) {
-      throw new NotFoundException(`User of id:${userId} not found`);
-    }
-    if (StatusArray.includes(status)) {
-      user.status = status;
-      // console.log('status:', status);
-    } else {
-      throw new BadRequestException(`User status: ${status} is not valid`);
-    }
-    const updatedUser = await this.userRepository.save(user);
-    this.chatEventsGateway.server.emit('status', updatedUser);
-    return updatedUser;
-  }
+  //   async updateUserStatus(userId: number, status: Status) {
+  //     const user = await this.userRepository
+  //       .createQueryBuilder('users')
+  //       .where('users.user_id=:userId', { userId })
+  //       .getOne();
+  //     if (!user) {
+  //       throw new NotFoundException(`User of id:${userId} not found`);
+  //     }
+  //     if (StatusArray.includes(status)) {
+  //       user.status = status;
+  //       // console.log('status:', status);
+  //     } else {
+  //       throw new BadRequestException(`User status: ${status} is not valid`);
+  //     }
+  //     const updatedUser = await this.userRepository.save(user);
+  //     this.chatEventsGateway.server.emit('status', updatedUser);
+  //     return updatedUser;
+  //   }
 }
