+++ backend/src/auth/strategies/ft.strategy.ts
@@ -3,13 +3,13 @@ import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-oauth2';
 import { lastValueFrom } from 'rxjs';
 import { HttpService } from '@nestjs/axios';
-import { IAuthService } from '../services/auth/auth.interface';
+import { IUserService } from 'src/users/services/user/user.interface';
 
 @Injectable()
 export class FtStrategy extends PassportStrategy(Strategy, 'ft') {
   constructor(
     private httpService: HttpService,
-    @Inject('AUTH_SERVICE') private readonly authService: IAuthService,
+    @Inject('USER_SERVICE') private readonly userService: IUserService,
   ) {
     super({
       authorizationURL: process.env.FT_AUTHORIZATION_URL,
@@ -42,7 +42,7 @@ export class FtStrategy extends PassportStrategy(Strategy, 'ft') {
       console.log('image_url: ', image_url);
       console.log('username: ', username);
       const userDetails = { intra_id, email, image_url, username };
-      return this.authService.validateUser(userDetails);
+      return this.userService.validateUser(userDetails);
     } catch (error) {
       console.error(error);
       throw new UnauthorizedException('Failed to authenticate by 42API');
